% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrR.seagrass-packages.R
\docType{package}
\name{arrR.seagrass}
\alias{arrR.seagrass}
\alias{arrR.seagrass-package}
\title{arrR.seagrass: Seagrass sub-processes of the arrR simulation model}
\description{
The goal of arrR.seagrass  is to provide all sub-processes related to the grid-based 
simulation environment of the arrR simulation model. This includes all procceses 
related to seagrass growth and the water column i.e., all  sub-processes that are 
simulated grid-based.
}
\author{
\strong{Maintainer}: Maximilian H.K. Hesselbarth \email{mhk.hesselbarth@gmail.com} (\href{https://orcid.org/0000-0003-1125-9918}{ORCID})

Authors:
\itemize{
  \item Kenzo E. Esquivel \email{kesquivel@berkeley.edu} (\href{https://orcid.org/0000-0002-1326-289X}{ORCID})
  \item Katrina S. Munsterman \email{kmunster@umich.edu} (\href{https://orcid.org/0000-0002-9303-2049}{ORCID})
}

Other contributors:
\itemize{
  \item Jacob E. Allgeier \email{jeallg@umich.edu} (\href{https://orcid.org/0000-0002-9005-6432}{ORCID}) [research team head]
}

}
